requirements.txt
alembic==1.10.3
anyio==3.6.2
attrs==22.2.0
black==23.3.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.95.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
install==1.3.5
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pandas==2.0.0
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pydantic==1.10.7
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
uvicorn==0.21.1
